+++ packages/typespec-ts/src/lib.ts
@@ -94,7 +94,15 @@ export const RLCOptionsSchema: JSONSchemaType<EmitterOptions> = {
     clearOutputFolder: { type: "boolean", nullable: true },
     ignorePropertyNameNormalize: { type: "boolean", nullable: true },
     ignoreEnumMemberNameNormalize: { type: "boolean", nullable: true },
-    compatibilityQueryMultiFormat: { type: "boolean", nullable: true }
+    compatibilityQueryMultiFormat: { type: "boolean", nullable: true },
+    typespecTitleMap: {
+      type: "object",
+      additionalProperties: {
+        type: "string"
+      },
+      required: [],
+      nullable: true
+    }
   },
   required: []
 };