+++ packages/typespec-ts/src/modular/emitModels.ts
@@ -39,14 +39,13 @@ import {
 
 import { SdkContext } from "../utils/interfaces.js";
 import { addDeclaration } from "../framework/declaration.js";
-import { addImportBySymbol } from "../utils/importHelper.js";
 import { buildModelDeserializer } from "./serialization/buildDeserializerFunction.js";
 import { buildModelSerializer } from "./serialization/buildSerializerFunction.js";
 import { extractPagedMetadataNested } from "../utils/operationUtil.js";
 import path from "path";
 import { refkey } from "../framework/refkey.js";
 import { useContext } from "../contextManager.js";
-import { isMetadata } from "@typespec/http";
+import { isMetadata, isOrExtendsHttpFile } from "@typespec/http";
 import {
   isAzureCoreErrorType,
   isAzureCoreLroType
@@ -60,6 +59,8 @@ import {
   normalizeModelPropertyName
 } from "./type-expressions/get-type-expression.js";
 import { emitQueue } from "../framework/hooks/sdkTypes.js";
+import { resolveReference } from "../framework/reference.js";
+import { MultipartHelpers } from "./static-helpers-metadata.js";
 
 type InterfaceStructure = OptionalKind<InterfaceDeclarationStructure> & {
   extends?: string[];
@@ -81,29 +82,19 @@ function isGenerableType(
     type.kind === "union" ||
     type.kind === "dict" ||
     type.kind === "array" ||
-    type.kind === "nullable"
+    (type.kind === "nullable" &&
+      isGenerableType(type.type) &&
+      Boolean(type.name) &&
+      !type.isGeneratedName)
   );
 }
 export function emitTypes(
   context: SdkContext,
   { sourceRoot }: { sourceRoot: string }
 ) {
-  const { sdkPackage } = context;
   const outputProject = useContext("outputProject");
 
-  const modelsFilePath = getModelsPath(sourceRoot);
   let sourceFile;
-  if (
-    emitQueue.size > 0 &&
-    (sdkPackage.models.length > 0 || sdkPackage.enums.length > 0)
-  ) {
-    sourceFile = outputProject.createSourceFile(modelsFilePath);
-    if (!sourceFile) {
-      throw new Error(`Failed to create source file at ${modelsFilePath}`);
-    }
-  } else {
-    return;
-  }
 
   for (const type of emitQueue) {
     if (!isGenerableType(type)) {
@@ -112,19 +103,31 @@ export function emitTypes(
     if (isAzureCoreLroType(type.__raw)) {
       continue;
     }
+
+    const namespaces = getModelNamespaces(context, type);
+    const filepath = getModelsPath(sourceRoot, namespaces);
+    sourceFile = outputProject.getSourceFile(filepath);
+    if (!sourceFile) {
+      sourceFile = outputProject.createSourceFile(filepath);
+    }
     emitType(context, type, sourceFile);
   }
 
-  if (
-    sourceFile.getInterfaces().length === 0 &&
-    sourceFile.getTypeAliases().length === 0 &&
-    sourceFile.getEnums().length === 0
-  ) {
-    sourceFile.delete();
-    return;
+  const modelFiles = outputProject.getSourceFiles(
+    sourceRoot + "/models/**/*.ts"
+  );
+  for (const modelFile of modelFiles) {
+    if (
+      modelFile.getInterfaces().length === 0 &&
+      modelFile.getTypeAliases().length === 0 &&
+      modelFile.getEnums().length === 0
+    ) {
+      modelFile.delete();
+      return;
+    }
   }
-  addImportBySymbol("serializeRecord", sourceFile);
-  return sourceFile;
+
+  return modelFiles;
 }
 
 function emitType(context: SdkContext, type: SdkType, sourceFile: SourceFile) {
@@ -132,11 +135,15 @@ function emitType(context: SdkContext, type: SdkType, sourceFile: SourceFile) {
     if (isAzureCoreErrorType(context.program, type.__raw)) {
       return;
     }
+    if (isOrExtendsHttpFile(context.program, type.__raw!)) {
+      return;
+    }
     if (
       !type.usage ||
       (type.usage !== undefined &&
         (type.usage & UsageFlags.Output) !== UsageFlags.Output &&
-        (type.usage & UsageFlags.Input) !== UsageFlags.Input)
+        (type.usage & UsageFlags.Input) !== UsageFlags.Input &&
+        (type.usage & UsageFlags.Exception) !== UsageFlags.Exception)
     ) {
       return;
     }
@@ -179,7 +186,9 @@ function emitType(context: SdkContext, type: SdkType, sourceFile: SourceFile) {
     const apiVersionEnumOnly = type.usage === UsageFlags.ApiVersionEnum;
     const inputUsage = (type.usage & UsageFlags.Input) === UsageFlags.Input;
     const outputUsage = (type.usage & UsageFlags.Output) === UsageFlags.Output;
-    if (!(inputUsage || outputUsage || apiVersionEnumOnly)) {
+    const exceptionUsage =
+      (type.usage & UsageFlags.Exception) === UsageFlags.Exception;
+    if (!(inputUsage || outputUsage || apiVersionEnumOnly || exceptionUsage)) {
       return;
     }
     const [enumType, knownValuesEnum] = buildEnumTypes(context, type);
@@ -208,6 +217,9 @@ function emitType(context: SdkContext, type: SdkType, sourceFile: SourceFile) {
     addSerializationFunctions(context, type, sourceFile);
   } else if (type.kind === "array") {
     addSerializationFunctions(context, type, sourceFile);
+  } else if (type.kind === "nullable") {
+    const nullableType = buildNullableType(context, type);
+    addDeclaration(sourceFile, nullableType, type);
   }
 }
 
@@ -221,8 +233,53 @@ export function getApiVersionEnum(context: SdkContext) {
   return apiVersionEnum;
 }
 
-export function getModelsPath(sourceRoot: string): string {
-  return path.join(...[sourceRoot, "models", `models.ts`]);
+export function getModelsPath(
+  sourceRoot: string,
+  modelNamespace: string[] = []
+): string {
+  return path.join(
+    ...[
+      sourceRoot,
+      "models",
+      ...modelNamespace.map((n) => normalizeName(n, NameType.File)),
+      `models.ts`
+    ]
+  );
+}
+
+export function getModelNamespaces(
+  context: SdkContext,
+  model: SdkType
+): string[] {
+  const rootNamespace = context.sdkPackage.rootNamespace.split(".");
+  if (
+    model.kind === "model" ||
+    model.kind === "enum" ||
+    model.kind === "union"
+  ) {
+    if (
+      model.clientNamespace.startsWith("Azure.ResourceManager") ||
+      model.clientNamespace.startsWith("Azure.Core") ||
+      model.crossLanguageDefinitionId.startsWith("TypeSpec.Rest.Resource") ||
+      model.crossLanguageDefinitionId === "TypeSpec.Http.File" // filter out the TypeSpec.Http.File model similar like what java does here https://github.com/microsoft/typespec/blob/main/packages/http-client-java/emitter/src/code-model-builder.ts#L2589
+    ) {
+      return [];
+    }
+    const segments = model.clientNamespace.split(".");
+    if (segments.length > rootNamespace.length) {
+      while (segments[0] === rootNamespace[0]) {
+        segments.shift();
+        rootNamespace.shift();
+      }
+      return segments;
+    }
+    return [];
+  } else if (model.kind === "array" || model.kind === "dict") {
+    return getModelNamespaces(context, model.valueType);
+  } else if (model.kind === "nullable") {
+    return getModelNamespaces(context, model.type);
+  }
+  return [];
 }
 
 function addSerializationFunctions(
@@ -283,6 +340,19 @@ function buildUnionType(
   unionDeclaration.docs = [type.doc ?? `Alias for ${unionDeclaration.name}`];
 
   return unionDeclaration;
+}
+
+function buildNullableType(context: SdkContext, type: SdkNullableType) {
+  const nullableDeclaration: TypeAliasDeclarationStructure = {
+    kind: StructureKind.TypeAlias,
+    name: normalizeModelName(context, type),
+    isExported: true,
+    type: getTypeExpression(context, type.type) + " | null"
+  };
+  nullableDeclaration.docs = [
+    type.doc ?? `Alias for ${nullableDeclaration.name}`
+  ];
+  return nullableDeclaration;
 }
 
 export function buildEnumTypes(
@@ -460,7 +530,8 @@ export function normalizeModelName(
     | SdkEnumType
     | SdkUnionType
     | SdkArrayType
-    | SdkDictionaryType,
+    | SdkDictionaryType
+    | SdkNullableType,
   nameType: NameType = NameType.Interface,
   skipPolymorphicUnionSuffix = false
 ): string {
@@ -472,6 +543,10 @@ export function normalizeModelName(
       type.valueType as any,
       nameType
     )}>`;
+  }
+  // TODO see https://github.com/Azure/typespec-azure/issues/2125
+  if (type.kind === "nullable") {
+    return normalizeName(type.name, nameType, true);
   }
   if (type.kind !== "model" && type.kind !== "enum" && type.kind !== "union") {
     return getTypeExpression(context, type);
@@ -551,10 +626,52 @@ function buildModelProperty(
       target: NoTarget
     });
   }
+
+  let typeExpression: string;
+  if (property.kind === "property" && property.isMultipartFileInput) {
+    const multipartOptions = property.multipartOptions;
+    typeExpression = "{";
+    typeExpression += `contents: ${resolveReference(MultipartHelpers.FileContents)};`;
+
+    const isContentTypeOptional =
+      multipartOptions?.contentType === undefined ||
+      multipartOptions.contentType.optional ||
+      multipartOptions.defaultContentTypes.length > 0;
+    const isFilenameOptional =
+      multipartOptions?.filename === undefined ||
+      multipartOptions.filename.optional;
+
+    const contentTypeType = multipartOptions?.contentType
+      ? getTypeExpression(context, multipartOptions.contentType.type)
+      : "string";
+    const filenameType = multipartOptions?.filename
+      ? getTypeExpression(context, multipartOptions.filename.type)
+      : "string";
+
+    typeExpression += `contentType${isContentTypeOptional ? "?" : ""}: ${contentTypeType};`;
+    typeExpression += `filename${isFilenameOptional ? "?" : ""}: ${filenameType};`;
+
+    typeExpression += "}";
+
+    if (isContentTypeOptional && isFilenameOptional) {
+      // Allow passing content directly if both filename and content type are optional
+      typeExpression = `(${resolveReference(MultipartHelpers.FileContents)}) | ${typeExpression}`;
+    } else {
+      // If either one is required, still accept File at the top level since it requires a filename
+      typeExpression = `File | ${typeExpression}`;
+    }
+
+    if (property.type.kind === "array") {
+      typeExpression = `Array<${typeExpression}>`;
+    }
+  } else {
+    typeExpression = getTypeExpression(context, property.type);
+  }
+
   const propertyStructure: PropertySignatureStructure = {
     kind: StructureKind.PropertySignature,
     name: normalizedPropName,
-    type: getTypeExpression(context, property.type),
+    type: typeExpression,
     hasQuestionToken: property.optional,
     isReadonly: isReadOnly(property as SdkBodyModelPropertyType)
   };