+++ packages/typespec-ts/src/modular/helpers/operationHelpers.ts
@@ -168,7 +168,7 @@ export function getDeserializePrivateFunction(
   );
   statements.push(
     `if(!expectedStatuses.includes(result.status)){`,
-    `throw ${createRestErrorReference}(result);`,
+    `${getExceptionThrowStatement(context, operation)}`,
     "}"
   );
   const deserializedType = isLroOnly
@@ -186,7 +186,7 @@ export function getDeserializePrivateFunction(
   if (isLroOnly && lroSubPath) {
     statements.push(
       `if(${deserializedRoot.split(".").join("?.")} === undefined) {
-        throw createRestError(\`Expected a result in the response at position "${deserializedRoot}"\`, result);
+        throw ${createRestErrorReference}(\`Expected a result in the response at position "${deserializedRoot}"\`, result);
       }
       `
     );
@@ -226,6 +226,103 @@ export function getDeserializePrivateFunction(
     ...functionStatement,
     statements
   };
+}
+
+interface ExceptionThrowDetail {
+  start: number;
+  end?: number;
+  deserializer: string;
+}
+
+interface OperationExceptionDetails {
+  customized: ExceptionThrowDetail[];
+  defaultDeserializer?: string;
+}
+
+function getExceptionDetails(
+  context: SdkContext,
+  operation: ServiceOperation
+): OperationExceptionDetails {
+  const customized: ExceptionThrowDetail[] = [];
+  let defaultDeserializer: string | undefined;
+  for (const exception of operation.operation.exceptions) {
+    if (!exception.type) {
+      continue;
+    }
+    const statusCode = exception.statusCodes;
+    const deserializeFunctionName = buildModelDeserializer(
+      context,
+      exception.type,
+      false,
+      true
+    );
+    if (
+      !deserializeFunctionName ||
+      typeof deserializeFunctionName !== "string"
+    ) {
+      continue;
+    }
+    if (statusCode === "*") {
+      defaultDeserializer = deserializeFunctionName;
+    } else if (typeof statusCode === "number") {
+      customized.push({
+        start: statusCode,
+        deserializer: deserializeFunctionName
+      });
+    } else {
+      customized.push({
+        start: statusCode.start,
+        end: statusCode.end,
+        deserializer: deserializeFunctionName
+      });
+    }
+  }
+  return { customized, defaultDeserializer };
+}
+
+function getExceptionThrowStatement(
+  context: SdkContext,
+  operation: ServiceOperation
+) {
+  const statements = [];
+  const createRestErrorReference = resolveReference(
+    useDependencies().createRestError
+  );
+  const { customized, defaultDeserializer } = getExceptionDetails(
+    context,
+    operation
+  );
+  if (customized.length > 0) {
+    statements.push(`const error = ${createRestErrorReference}(result);`);
+    statements.push(`const statusCode = Number.parseInt(result.status);`);
+    const stats: string[] = customized.map((exception) => {
+      if (exception.end) {
+        return `if(statusCode >= ${exception.start} && statusCode <= ${exception.end}) {
+              error.details = ${exception.deserializer}(result.body);
+          }`;
+      } else {
+        return `if(statusCode === ${exception.start}) {
+             error.details = ${exception.deserializer}(result.body);
+          }`;
+      }
+    });
+    statements.push(stats.join("\nelse "));
+    if (defaultDeserializer) {
+      statements.push(`else {
+        error.details = ${defaultDeserializer}(result.body);
+      }`);
+    }
+    statements.push("throw error;");
+  } else {
+    if (defaultDeserializer) {
+      statements.push(`const error = ${createRestErrorReference}(result);
+      error.details = ${defaultDeserializer}(result.body);`);
+      statements.push("throw error;");
+    } else {
+      statements.push(`throw ${createRestErrorReference}(result);`);
+    }
+  }
+  return statements.join("\n");
 }
 
 function getOptionalParamsName(
@@ -492,20 +589,14 @@ function getPagingOnlyOperationFunction(
 
   const statements: string[] = [];
   const options = [];
-  // TODO pending tcgc issue to fix https://github.com/Azure/typespec-azure/issues/1985
-  const itemName =
-    operation.response.resultPath !== "" && operation.response.resultPath
-      ? operation.response.resultPath
-      : operation.__raw_paged_metadata?.itemsSegments
-        ? operation.__raw_paged_metadata.itemsSegments[0]
-        : undefined;
+  // TODO: follow up on https://github.com/Azure/typespec-azure/issues/2103
+  const nextLinkName = operation.nextLinkPath;
+  const itemName = operation.response.resultPath;
   if (itemName) {
     options.push(`itemName: "${itemName}"`);
   }
-  if (operation.__raw_paged_metadata?.nextLinkSegments) {
-    options.push(
-      `nextLinkName: "${operation.__raw_paged_metadata.nextLinkSegments[0]}"`
-    );
+  if (nextLinkName) {
+    options.push(`nextLinkName: "${nextLinkName}"`);
   }
   statements.push(
     `return ${buildPagedAsyncIteratorReference}(
@@ -795,7 +886,7 @@ function getRequired(context: SdkContext, param: SdkModelPropertyType) {
   const serializedName = getPropertySerializedName(param);
   const clientValue = `${param.onClient ? "context." : ""}${param.name}`;
   if (param.type.kind === "model") {
-    const { propertiesStr } = getRequestModelMapping(
+    const propertiesStr = getRequestModelMapping(
       context,
       { ...param.type, optional: param.optional },
       clientValue
@@ -834,15 +925,12 @@ function getOptional(
   const serializedName = getPropertySerializedName(param);
   const paramName = `${param.onClient ? "context." : `${optionalParamName}?.`}${param.name}`;
   if (param.type.kind === "model") {
-    const { propertiesStr, directAssignment } = getRequestModelMapping(
+    const propertiesStr = getRequestModelMapping(
       context,
       { ...param.type, optional: param.optional },
       paramName + "?."
     );
-    const serializeContent =
-      directAssignment === true
-        ? propertiesStr.join(",")
-        : `{${propertiesStr.join(",")}}`;
+    const serializeContent = `{${propertiesStr.join(",")}}`;
     return `"${serializedName}": ${serializeContent}`;
   }
   return `"${serializedName}": ${serializeRequestValue(
@@ -948,76 +1036,92 @@ function getNullableCheck(name: string, type: SdkType) {
   return `${name} === null ? null :`;
 }
 
-/**
- *
- * This function helps translating an HLC request to RLC request,
- * extracting properties from body and headers and building the RLC response object
- */
-interface RequestModelMappingResult {
-  propertiesStr: string[];
-  directAssignment?: boolean;
+export function getSerializationExpression(
+  context: SdkContext,
+  property: SdkModelPropertyType,
+  propertyPath: string
+): string {
+  const dot = propertyPath.endsWith("?") ? "." : "";
+  const propertyPathWithDot = `${propertyPath ? `${propertyPath}${dot}` : `${dot}`}`;
+  const nullOrUndefinedPrefix = getPropertySerializationPrefix(
+    context,
+    property,
+    propertyPath
+  );
+
+  const propertyFullName = getPropertyFullName(
+    context,
+    property,
+    propertyPathWithDot
+  );
+  const serializeFunctionName = buildModelSerializer(
+    context,
+    getNullableValidType(property.type),
+    false,
+    true
+  );
+  if (serializeFunctionName) {
+    return `${nullOrUndefinedPrefix}${serializeFunctionName}(${propertyFullName})`;
+  } else if (isAzureCoreErrorType(context.program, property.type.__raw)) {
+    return `${nullOrUndefinedPrefix}${propertyFullName}`;
+  } else {
+    return serializeRequestValue(
+      context,
+      property.type,
+      propertyFullName,
+      !property.optional,
+      getEncodeForType(property.type)
+    );
+  }
 }
-export function getRequestModelMapping(
+
+export function getRequestModelProperties(
   context: SdkContext,
   modelPropertyType: SdkModelType & { optional?: boolean },
   propertyPath: string = "body"
-): RequestModelMappingResult {
-  const props: string[] = [];
+): Array<[string, string]> {
+  const props: [string, string][] = [];
   const allParents = getAllAncestors(modelPropertyType);
   const properties: SdkModelPropertyType[] =
     getAllProperties(modelPropertyType, allParents) ?? [];
   if (properties.length <= 0) {
-    return { propertiesStr: [] };
+    return [];
   }
   for (const property of properties) {
     if (property.kind === "property" && isReadOnly(property)) {
       continue;
     }
-    const dot = propertyPath.endsWith("?") ? "." : "";
-    const serializedName = getPropertySerializedName(property);
-    const propertyPathWithDot = `${propertyPath ? `${propertyPath}${dot}` : `${dot}`}`;
-    const nullOrUndefinedPrefix = getPropertySerializationPrefix(
-      context,
-      property,
-      propertyPath
-    );
 
-    const propertyFullName = getPropertyFullName(
-      context,
-      property,
-      propertyPathWithDot
-    );
-    const serializeFunctionName = buildModelSerializer(
-      context,
-      getNullableValidType(property.type),
-      false,
-      true
-    );
-    if (serializeFunctionName) {
-      props.push(
-        `"${serializedName}": ${nullOrUndefinedPrefix}${serializeFunctionName}(${propertyFullName})`
-      );
-    } else if (isAzureCoreErrorType(context.program, property.type.__raw)) {
-      props.push(
-        `"${serializedName}": ${nullOrUndefinedPrefix}${propertyFullName}`
-      );
-    } else {
-      const serializedValue = serializeRequestValue(
-        context,
-        property.type,
-        propertyFullName,
-        !property.optional,
-        getEncodeForType(property.type)
-      );
-      props.push(`"${serializedName}": ${serializedValue}`);
-    }
+    props.push([
+      getPropertySerializedName(property)!,
+      getSerializationExpression(context, property, propertyPath)
+    ]);
   }
 
-  return { propertiesStr: props };
+  return props;
+}
+
+/**
+ *
+ * This function helps translating an HLC request to RLC request,
+ * extracting properties from body and headers and building the RLC response object
+ */
+export function getRequestModelMapping(
+  context: SdkContext,
+  modelPropertyType: SdkModelType & { optional?: boolean },
+  propertyPath: string = "body"
+): string[] {
+  return getRequestModelProperties(
+    context,
+    modelPropertyType,
+    propertyPath
+  ).map(([name, value]) => `"${name}": ${value}`);
 }
 
 function getPropertySerializedName(property: SdkModelPropertyType) {
-  return property.kind !== "credential" && property.kind !== "method"
+  return property.kind !== "credential" &&
+    property.kind !== "method" &&
+    property.kind !== "endpoint"
     ? property.serializedName
     : property.name;
 }
@@ -1164,7 +1268,7 @@ export function serializeRequestValue(
         return `${clientValue} as any`;
       }
     case "model": // this is to build serialization logic for spread model types
-      return `{${getRequestModelMapping(context, type, "").propertiesStr.join(",")}}`;
+      return `{${getRequestModelMapping(context, type, "").join(",")}}`;
     case "nullable":
       return serializeRequestValue(
         context,